Simpler cache API
