+++ packages/backend/src/modules/update-monitor/DiscoveryCache.ts
@@ -10,12 +10,7 @@ export class DiscoveryCache implements DiscoveryCacheInterface {
     return record?.value
   }
 
-  async set(
-    key: string,
-    value: string,
-    chain: string,
-    blockNumber: number,
-  ): Promise<void> {
-    await this.db.discoveryCache.upsert({ key, value, chain, blockNumber })
+  async set(key: string, value: string): Promise<void> {
+    await this.db.discoveryCache.upsert({ key, value })
   }
 }
