+++ packages/database/src/discovery/discovery-cache/select.ts
@@ -3,6 +3,4 @@ import { DiscoveryCache } from '../../kysely/generated/types'
 export const selectDiscoveryCache = [
   'key',
   'value',
-  'chain',
-  'blockNumber',
 ] as const satisfies (keyof DiscoveryCache)[]
