+++ packages/discovery/src/discovery/provider/AllProviders.ts
@@ -6,10 +6,11 @@ import { providers } from 'ethers'
 import { DiscoveryChainConfig } from '../../config/types'
 import { getExplorerClient } from '../../utils/IEtherscanClient'
 import { BatchingAndCachingProvider } from './BatchingAndCachingProvider'
+import { DiscoveryCache } from './DiscoveryCache'
 import { HighLevelProvider } from './HighLevelProvider'
 import { IProvider, RawProviders } from './IProvider'
 import { LowLevelProvider } from './LowLevelProvider'
-import { DiscoveryCache, ReorgAwareCache } from './ReorgAwareCache'
+import { ReorgAwareCache } from './ReorgAwareCache'
 import { AllProviderStats, ProviderStats } from './Stats'
 import { getBlockNumberTwoProviders } from './getBlockNumberTwoProviders'
 import { MulticallClient } from './multicall/MulticallClient'
