+++ packages/discovery/src/discovery/provider/BatchingAndCachingProvider.ts
@@ -11,9 +11,10 @@ import { BigNumber, providers } from 'ethers'
 import { ContractSource } from '../../utils/IEtherscanClient'
 import { isRevert } from '../utils/isRevert'
 import { DebugTransactionCallResponse } from './DebugTransactionTrace'
+import { CacheEntry } from './DiscoveryCache'
 import { ContractDeployment, RawProviders } from './IProvider'
 import { LowLevelProvider } from './LowLevelProvider'
-import { CacheEntry, ReorgAwareCache } from './ReorgAwareCache'
+import { ReorgAwareCache } from './ReorgAwareCache'
 import { ProviderMeasurement, ProviderStats } from './Stats'
 import { MulticallClient } from './multicall/MulticallClient'
