+++ packages/discovery/src/discovery/provider/SQLiteCache.test.ts
@@ -10,10 +10,8 @@ describe('SQLiteCache', () => {
     withTemporaryFile(async (sqlCache, rqe) => {
       const key = 'key'
       const value = 'value'
-      const chain = 'ethereum'
-      const blockNumber = 1
 
-      await sqlCache.set(key, value, chain, blockNumber)
+      await sqlCache.set(key, value)
       const queriedValue = await sqlCache.get(key)
 
       const resultRaw = await rqe.query<CacheEntry[]>(
@@ -36,18 +34,13 @@ describe('SQLiteCache', () => {
   it('replaces old value in case of conflict', () =>
     withTemporaryFile(async (sqlCache, rqe) => {
       const key = 'key'
-
       const value = 'value'
-      const chain = 'ethereum'
-      const blockNumber = 1
 
       const newValue = 'newValue'
-      const newChain = 'arbitrum'
-      const newBlockNumber = 2
 
-      await sqlCache.set(key, value, chain, blockNumber)
+      await sqlCache.set(key, value)
 
-      await sqlCache.set(key, newValue, newChain, newBlockNumber)
+      await sqlCache.set(key, newValue)
 
       const resultRaw = await rqe.query<CacheEntry[]>(
         'SELECT * FROM cache WHERE key=$1',
@@ -61,8 +54,6 @@ describe('SQLiteCache', () => {
       expect(resultRaw.length).toEqual(1)
       expect(result.key).toEqual(key)
       expect(result.value).toEqual(newValue)
-      expect(result.blockNumber).toEqual(newBlockNumber)
-      expect(result.chain).toEqual(newChain)
     }))
 })
