+++ packages/discovery/src/discovery/provider/SQLiteCache.ts
@@ -1,6 +1,6 @@
 import { existsSync, mkdirSync } from 'fs'
 import sqlite3 from 'sqlite3' // SQLite Client
-import { DiscoveryCache } from './ReorgAwareCache'
+import { DiscoveryCache } from './DiscoveryCache'
 
 const DEFAULT_DATABASE_DIR = 'cache'
 const DEFAULT_DATABASE_FILENAME = 'discovery.sqlite'
@@ -22,8 +22,6 @@ export class SQLiteCache implements DiscoveryCache {
     await this.query(`
       CREATE TABLE IF NOT EXISTS cache (
         key TEXT PRIMARY KEY,
-        blockNumber INTEGER,
-        chain TEXT,
         value TEXT
       )
     `)
@@ -40,19 +38,14 @@ export class SQLiteCache implements DiscoveryCache {
     }
   }
 
-  async set(
-    key: string,
-    value: string,
-    chain: string,
-    blockNumber: number | undefined,
-  ): Promise<void> {
+  async set(key: string, value: string): Promise<void> {
     try {
       await this.query(
         `
-        INSERT INTO cache(key, value, chain, blockNumber) 
-        VALUES($1, $2, $3, $4) 
-        ON CONFLICT(key) DO UPDATE SET value=$2, chain=$3, blockNumber=$4`,
-        [key, value, chain, blockNumber],
+        INSERT INTO cache(key, value)
+        VALUES($1, $2) 
+        ON CONFLICT(key) DO UPDATE SET value=$2`,
+        [key, value],
       )
     } catch (error) {
       console.error('Error writing to cache', error)
