+++ packages/discovery/src/index.ts
@@ -40,7 +40,7 @@ export { SourceCodeService } from './discovery/source/SourceCodeService'
 export { normalizeDiffPath } from './discovery/utils/normalizeDiffPath'
 export { sortBySeverity } from './discovery/utils/sortDiffs'
 export type { HashedChunks, HashedFileContent } from './flatten/utils'
-export { type DiscoveryCache } from './discovery/provider/ReorgAwareCache'
+export { type DiscoveryCache } from './discovery/provider/DiscoveryCache'
 export { SQLiteCache } from './discovery/provider/SQLiteCache'
 export { NoCache } from './discovery/provider/NoCache'
 export {
